pyproject.toml
[project]
name = "pdf-qna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.7",
    "langchain-openai>=0.2.8",
    "langchain-text-splitters>=0.3.2",
    "openai>=1.54.4",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "rank-bm25>=0.2.2",
    "langchain-community>=0.3.7",
    "faiss-cpu>=1.9.0",
    "streamlit>=1.40.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
